+++ sdk/resourcemanager/azure-resourcemanager-resources/src/main/java/com/azure/resourcemanager/resources/models/Tenants.java
@@ -5,12 +5,11 @@ package com.azure.resourcemanager.resources.models;
 
 import com.azure.core.annotation.Fluent;
 import com.azure.resourcemanager.resources.fluentcore.collection.SupportsListing;
-import com.azure.resourcemanager.resources.fluent.models.TenantIdDescriptionInner;
 
 /**
  * Entry point to tenant management API.
  */
 @Fluent
 public interface Tenants extends
-        SupportsListing<TenantIdDescriptionInner> {
+        SupportsListing<Tenant> {
 }