+++ sdk/servicebus/azure-messaging-servicebus/src/main/java/com/azure/messaging/servicebus/implementation/MessageUtils.java
@@ -9,10 +9,8 @@ import com.azure.core.amqp.implementation.AmqpConstants;
 import com.azure.core.amqp.implementation.TracerProvider;
 import com.azure.core.util.Context;
 import com.azure.core.util.CoreUtils;
-import com.azure.core.util.logging.ClientLogger;
 import com.azure.core.util.tracing.ProcessKind;
 import com.azure.messaging.servicebus.ServiceBusMessage;
-import com.azure.messaging.servicebus.ServiceBusReceivedMessage;
 import com.azure.messaging.servicebus.ServiceBusTransactionContext;
 import org.apache.qpid.proton.amqp.Binary;
 import org.apache.qpid.proton.amqp.Symbol;
@@ -25,15 +23,12 @@ import org.apache.qpid.proton.amqp.transport.DeliveryState;
 import org.apache.qpid.proton.amqp.transport.ErrorCondition;
 import reactor.core.publisher.Signal;
 
-import java.io.Closeable;
-import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.time.Duration;
 import java.time.Instant;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 import java.util.HashMap;
-import java.util.Locale;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
@@ -42,8 +37,6 @@ import static com.azure.core.util.tracing.Tracer.AZ_TRACING_NAMESPACE_KEY;
 import static com.azure.core.util.tracing.Tracer.DIAGNOSTIC_ID_KEY;
 import static com.azure.core.util.tracing.Tracer.ENTITY_PATH_KEY;
 import static com.azure.core.util.tracing.Tracer.HOST_NAME_KEY;
-import static com.azure.core.util.tracing.Tracer.MESSAGE_ENQUEUED_TIME;
-import static com.azure.core.util.tracing.Tracer.SCOPE_KEY;
 import static com.azure.core.util.tracing.Tracer.SPAN_CONTEXT_KEY;
 import static com.azure.messaging.servicebus.implementation.ServiceBusConstants.AZ_TRACING_SERVICE_NAME;
 import static com.azure.messaging.servicebus.implementation.ServiceBusConstants.AZ_TRACING_NAMESPACE_VALUE;
@@ -328,52 +321,5 @@ public final class MessageUtils {
             }
         }
         return serviceBusMessage;
-    }
-
-    /*
-     * Starts a new process tracing span and attaches the returned context to the ServiceBusReceivedMessage object for
-     * users.
-     */
-    public static Context startProcessTracingSpan(ServiceBusReceivedMessage receivedMessage,
-        String hostname, String entityPath, TracerProvider tracerProvider, ProcessKind processKind) {
-        Object diagnosticId = receivedMessage.getApplicationProperties().get(DIAGNOSTIC_ID_KEY);
-        if (diagnosticId == null || !tracerProvider.isEnabled()) {
-            return Context.NONE;
-        }
-
-        Context spanContext = tracerProvider.extractContext(diagnosticId.toString(), Context.NONE)
-            .addData(ENTITY_PATH_KEY, entityPath)
-            .addData(HOST_NAME_KEY, hostname)
-            .addData(AZ_TRACING_NAMESPACE_KEY, AZ_TRACING_NAMESPACE_VALUE);
-        spanContext = receivedMessage.getEnqueuedTime() == null
-            ? spanContext
-            : spanContext.addData(MESSAGE_ENQUEUED_TIME, receivedMessage.getEnqueuedTime().toEpochSecond());
-        return tracerProvider.startSpan(AZ_TRACING_SERVICE_NAME, spanContext, processKind);
-    }
-
-    /*
-     * Ends the process tracing span and the scope of that span.
-     */
-    public static void endProcessTracingSpan(Context processSpanContext, Signal<Void> signal,
-        TracerProvider tracerProvider, ClientLogger logger) {
-        if (processSpanContext != null) {
-            Optional<Object> spanScope = processSpanContext.getData(SCOPE_KEY);
-            // Disposes of the scope when the trace span closes.
-            if (tracerProvider.isEnabled() && spanScope.isPresent()) {
-                if (spanScope.get() instanceof Closeable) {
-                    Closeable close = (Closeable) spanScope.get();
-                    try {
-                        close.close();
-                        tracerProvider.endSpan(processSpanContext, signal);
-                    } catch (IOException ioException) {
-                        logger.error(Messages.MESSAGE_PROCESSOR_RUN_END, ioException);
-                    }
-
-                } else {
-                    logger.warning(String.format(Locale.US,
-                        Messages.PROCESS_SPAN_SCOPE_TYPE_ERROR, spanScope.getClass()));
-                }
-            }
-        }
     }
 }