+++ sdk/resourcemanager/azure-resourcemanager-resources/CHANGELOG.md
@@ -1,7 +1,9 @@
 # Release History
 
-## 2.1.0-beta.1 (Unreleased)
+## 2.1.0 (2020-11-24)
 
+- Changed `ProviderRegistrationPolicy` using `Providers` endpoint directly.
+- Changed `Tenants::list` to return new interface `Tenant` instead of `TenantIdDescriptionInner`.
 
 ## 2.0.0 (2020-10-19)